refactor: Merge schema for API to types
