+++ src/components/Chart.tsx
@@ -1,7 +1,7 @@
 import {useQuery} from 'react-query';
 import ApexChart from 'react-apexcharts';
 import {IMAGES} from '../constants/images';
-import {CoinChartSchema} from '../api/schema/coinSchema';
+import {CoinChartSchema} from '../types';
 import {getCoinChartData} from '../api/coin';
 import {ChartProps} from '../types';
 import {useRecoilValue} from 'recoil';
